+++ src/ValidationErrorCollection.php
@@ -1,28 +1,29 @@
 <?php
 /**
- * Bright Nucleus Values.
+ * Bright Nucleus Options Store.
  *
- * Manipulate sanitizable and validatable values.
+ * Abstract options store that allows for exchangeable persistence mechanisms.
  *
- * @package   BrightNucleus\Values
+ * @package   BrightNucleus\OptionsStore
  * @author    Alain Schlesser <alain.schlesser@gmail.com>
  * @license   MIT
  * @link      https://www.brightnucleus.com/
  * @copyright 2017 Alain Schlesser, Bright Nucleus
  */
 
-namespace BrightNucleus\Values;
+namespace BrightNucleus\OptionsStore;
 
 use Doctrine\Common\Collections\ArrayCollection;
 
 /**
  * Class ValidationErrorCollection.
  *
- * @since   0.1.1
+ * @since   0.1.11
  *
  * @package BrightNucleus\Values
  * @author  Alain Schlesser <alain.schlesser@gmail.com>
  */
-class ValidationErrorCollection extends ArrayCollection {
+class ValidationErrorCollection extends ArrayCollection
+{
 
 }